apps/openchallenges/edam-etl/pyproject.toml
[tool.poetry]
name = "openchallenges-edam-etl"
version = "0.1.0"
description = "OpenChallenges EDAM ETL"
authors = ["Thomas Schaffter <thomas.schaffter@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "3.12.0"
pandas = "2.2.1"
requests = "2.31.0"
regex = "2023.12.25"
mariadb = "^1.1.10"
mysql-connector = "^2.2.9"
sqlalchemy = "^2.0.30"

[tool.poetry.group.dev.dependencies]

[tool.poetry.group.prod.dependencies]

[tool.poetry.group.test.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
